fixed vocabulary row height
