+++ worddeposit/XIBs/VocabularyWordBubbleView.xib
@@ -15,7 +15,7 @@
             <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
             <subviews>
                 <stackView opaque="NO" contentMode="scaleToFill" axis="vertical" alignment="center" spacing="20" translatesAutoresizingMaskIntoConstraints="NO" id="3we-W3-ImE">
-                    <rect key="frame" x="20" y="20" width="367" height="261.5"/>
+                    <rect key="frame" x="20" y="20" width="367" height="245.5"/>
                     <subviews>
                         <imageView clipsSubviews="YES" userInteractionEnabled="NO" contentMode="scaleAspectFit" horizontalHuggingPriority="251" verticalHuggingPriority="251" translatesAutoresizingMaskIntoConstraints="NO" id="9WQ-FB-tWo">
                             <rect key="frame" x="113.5" y="0.0" width="140" height="140"/>
@@ -24,8 +24,8 @@
                                 <constraint firstAttribute="height" constant="140" id="tXI-Wb-0Ef"/>
                             </constraints>
                         </imageView>
-                        <stackView opaque="NO" contentMode="scaleToFill" axis="vertical" spacing="20" translatesAutoresizingMaskIntoConstraints="NO" id="DZ3-bt-Q7Z">
-                            <rect key="frame" x="140" y="160" width="87.5" height="101.5"/>
+                        <stackView opaque="NO" contentMode="scaleToFill" axis="vertical" spacing="12" translatesAutoresizingMaskIntoConstraints="NO" id="DZ3-bt-Q7Z">
+                            <rect key="frame" x="140" y="160" width="87.5" height="85.5"/>
                             <subviews>
                                 <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" text="Example" textAlignment="center" lineBreakMode="tailTruncation" numberOfLines="0" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="DAV-Hc-f8k" customClass="VocabularyWordBubbleLabel" customModule="worddeposit" customModuleProvider="target">
                                     <rect key="frame" x="0.0" y="0.0" width="87.5" height="20.5"/>
@@ -34,13 +34,13 @@
                                     <nil key="highlightedColor"/>
                                 </label>
                                 <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="Translation" textAlignment="center" lineBreakMode="tailTruncation" numberOfLines="0" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="2VB-y7-FE8" customClass="VocabularyWordBubbleLabel" customModule="worddeposit" customModuleProvider="target">
-                                    <rect key="frame" x="0.0" y="40.5" width="87.5" height="20.5"/>
+                                    <rect key="frame" x="0.0" y="32.5" width="87.5" height="20.5"/>
                                     <fontDescription key="fontDescription" type="system" pointSize="17"/>
                                     <nil key="textColor"/>
                                     <nil key="highlightedColor"/>
                                 </label>
                                 <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="Description" textAlignment="center" lineBreakMode="tailTruncation" numberOfLines="0" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="mDh-si-2nk" customClass="VocabularyWordBubbleLabel" customModule="worddeposit" customModuleProvider="target">
-                                    <rect key="frame" x="0.0" y="81" width="87.5" height="20.5"/>
+                                    <rect key="frame" x="0.0" y="65" width="87.5" height="20.5"/>
                                     <fontDescription key="fontDescription" type="system" pointSize="17"/>
                                     <nil key="textColor"/>
                                     <nil key="highlightedColor"/>
